added cors config
