+++ src/main/resources/application.properties
@@ -35,6 +35,9 @@ upload.static-path=classpath:static
 # URL that will serve static content
 upload.static-serve=http://localhost:3000/static
 
+# Frontend Url
+frontend.url = http://localhost:3000
+
 # Due to a problem with Hibernate, which is using a deprecated property. This should be removed when fixed
 # See https://github.com/spring-projects/spring-data-jpa/issues/2717 for more information
 spring.jpa.properties.jakarta.persistence.sharedCache.mode=UNSPECIFIED
